% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_finder.R
\name{PiCal}
\alias{PiCal}
\title{Calculate the value of Pi}
\usage{
PiCal(sample.size)
}
\arguments{
\item{sample.size}{the sample size used to calculate Pi}
}
\value{
Pi, sample, sample indicator
}
\description{
This function used coordinates generated from two uniform distributions and count how many of
them falled into a circle to calculate the value of Pi
}
\examples{
PiCal(30)
PiCal(300)

}
